pyproject.toml
[project]
name = "spend-rail"
version = "0.1.0"
description = "FastAPI image categorization service with Gemini AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "python-multipart>=0.0.9",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
    "aiofiles>=24.0.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.pytest.ini_options]
pythonpath = ["src"]

